requirements.txt
flask
streamlit
streamlit-folium
folium
pandas
pillow
requests
gunicorn
